update: use m calculator
